Componentize sources
