+++ src/components/footer.js
@@ -1,12 +1,18 @@
-import React from 'react';
+import React, {Fragment} from 'react';
 import {SectionWrapper} from './common';
 import {Typography} from '@material-ui/core';
+
+export function Colophon() {
+  return (
+    <Fragment>&copy; {new Date().getFullYear()} Really Awesome Doings</Fragment>
+  );
+}
 
 export default function Footer() {
   return (
     <SectionWrapper>
       <Typography color="textSecondary">
-        &copy; {new Date().getFullYear()} Really Awesome Doings
+        <Colophon />
       </Typography>
     </SectionWrapper>
   );
